Add airplane mode state observer and modifier
